% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3TimeMap-Rd.R, R/d3TimeMap.R
\name{d3TimeMap-shiny}
\alias{d3TimeMap-shiny}
\alias{d3TimeMapOutput}
\alias{renderD3TimeMap}
\title{Shiny bindings for d3TimeMap widgets}
\usage{
d3TimeMapOutput(outputId, width = "100\%", height = "600px")

renderD3TimeMap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a d3TimeMap scatter plot.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}

\item{outputId}{Output variable to read from}

\item{width,height}{The width and height of the map (see
\link[htmlwidgets]{shinyWidgetOutput})}

\item{expr}{An expression that generates a \code{\link{d3TimeMap}} object}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using d3TimeMap widgets within Shiny
applications and interactive Rmd documents.

Use \code{d3TimeMapOutput} to create a UI element, and \code{renderD3TimeMap}
to render the timemap.
}

